requirements.txt
apache-airflow-providers-apache-kafka
apache-airflow-providers-postgres
kafka-python
pyspark
pyhive
psycopg2-binary
celery
redis
requests
pandas
pyarrow
fastparquet
great_expectations
fuzzywuzzy[speedup]
textblob
nltk